# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sosyal Medya API",
        "description": "backend",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/logs": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Sistem logları",
                "operationId": "d520fb1bca392da177d66c23b0b8915d",
                "responses": {
                    "200": {
                        "description": "Tüm aktiviteler",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 123
                                                    },
                                                    "log_name": {
                                                        "type": "string",
                                                        "example": "post"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Post created by user Ali"
                                                    },
                                                    "causer": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": {
        "1": {
            "sanctum": []
        }
    }
}